requirements.txt
Flask
gunicorn
Pillow
opencv-python
torch
torchvision
mediapipe
transformers
diffusers
safetensors
flask-cors
accelerate